Updating

+++ kifi-backend/modules/curator/test/com/keepit/curator/SeedIngestionCommanderTest.scala
@@ -207,7 +207,7 @@ class SeedIngestionCommanderTest extends Specification with CuratorTestInjector 
 
         db.readOnlyMaster { implicit session =>
           val seedItem1: Option[RawSeedItem] = seedItemRepo.getByUriIdAndUserId(user1Keeps.head.uriId, user1)
-          seedItem1.get.priorScore === Some(0.795.toFloat)
+          seedItem1.get.priorScore === Some(0.795f)
           seedItem1.get.userId === Some(Id[User](42))
           seedItem1.get.uriId === Id[NormalizedURI](1)
